requirements.txt
asgiref==3.7.2
Django==4.2.6
Faker==19.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
typing_extensions==4.8.0
